+++ tests/fixtures/accessibles/accessible.py
@@ -22,6 +22,14 @@ class SomeType:
     def simple_staticmethod(x: int) -> float:
         return float(x * 3)
 
+    async def simple_async_method(self, x: int) -> float:
+        return self._y * x * self._x
+
+
 
 def simple_function(z: float) -> float:
+    return z
+
+
+async def simple_async_function(z: float) -> float:
     return z